Move tests to separate folder and fix python packaging
