+++ Test-Dockerfile
@@ -37,6 +37,8 @@ RUN pip install -r /tmp/requirements.txt
 RUN mkdir -p /src
 
 COPY src/ /src/
+COPY tests/ /tests/
 
-WORKDIR /src
+WORKDIR tests/
+
 CMD /wait && python -m pytest
