+++ tests/e2e/test_api.py
@@ -1,8 +1,8 @@
 from flask.testing import FlaskClient
 import pytest
 from werkzeug.test import TestResponse
-from movies.entrypoints.flask_app import app
-from movies.adapters.sql_alchemy_repository import SqlAlchemyRepository
+from src.movies.entrypoints.flask_app import app
+from src.movies.adapters.sql_alchemy_repository import SqlAlchemyRepository
 
 # -- helpers
